fix(ai-assistant): exclude standalone loader from app bundle
